requirements.txt
asgiref==3.7.2
click==8.1.6
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
joblib==1.3.1
Markdown==3.4.3
nltk==3.8.1
numpy==1.25.1
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
scikit-learn==1.3.0
scipy==1.11.1
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.65.0
tzdata==2023.3
Unidecode==1.3.6
